Add files via upload
The index.html file is the basic frame deaign for UI appearance. The script.js being the logical structure for operations of the zkBTC-E minting simulation and the styles.css the binding glue for an aesthetic appeal. 

The Bit-Earth Labs Grail Pro Vault zkBTC-E Minter.html is the 'plug and play' compilation of all three for visualisation and experience of the demo in an integrated environment. It serves to give a more understandable step by step layout of how the authentication, minting and distribution of the Bitcoin-Energy Standard chain agnostic token works. Step 1 being the simulation of the IoT-Oracle Network-Grail Pro Vault interaction in a visualised manner; Step 2 being the prosumer experience simulation of how they can either bridge the zkBTC-E tokens to their wallets or burn them for an equivalent of other digital assets backing them prepaid by the consumer.